Adds original sampfreq function
Adds the sampfreq function from the original WFDB software package. This version will display the equivalent of using the -a option from the original sampfreq command since this displays the most information and the -H option can be reproduced using the output of the -a option. The -h option does not appear to do anything in the original source code.