Adds original ann2rr function
Adds the ann2rr function from the original WFDB software package. This version sets the -A option as the default in place of -c. Both the -f and -t options are conserved which specify the desired start and stop time. The -h option can be found using help(wfdb.ann2rr). The -i option is conserved except for the t parameter and specified number of decimal places. All the other options were not conserved since they can be determined using the current framework.