Converts relative imports to absolute
Converts all relative imports to absolute as per PEP8 guidelines. According to documentation... Absolute imports are recommended, as they are usually more readable and tend to be better behaved (or at least give better error messages) if the import system is incorrectly configured (such as when a directory inside a package ends up on sys.path). I actually encountered the previous error on #236 and had a tough time tracking down the error which prompted me to make this switch.