Fixes return_res error #230
Fixes error when using the return_res parameter which can sometimes modify the format in which the data from file is read. This oftentimes creates illogical arrays which can't be recovered into the proper values or even the proper shape. Fortunately, the removal of this does not significantly affect the runtime of the function even for large arrays. Fixes #230.