Produces WAV file from WFDB format
Introduces the conversion of WFDB files to WAV format. All data was written to the WAV file using struct.pack() and numpy.tofile() so no outside packages were introduced. This implementation conserves the -h option which can now be used with help(wfdb.mit2wav). Further, the -o option, which is used to specify their desired WAV file name, is also conserved in the form of the parameter output_filename. Finally, the -n option is conserved which is used to specify whether or not to write an accompanying header file. This is now in the parameter write_header.