Lazily import matplotlib.
A few functions (wfdb.plot_items, wfdb.plot.plot.create_figure,
wfdb.processing.Comparitor.plot) can be used to display graphical
plots using matplotlib.

When the matplotlib module is imported, it does a lot of non-trivial
initialization, which is slow and will sometimes fail in unpredictable
ways (depending on the environment and depending on whether other
nominally-optional dependencies, such as Tk, are installed.)

It's rather absurd for the behavior of importing wfdb to be dependent
on existence or non-existence of a window system, so only import
matplotlib in the few functions that actually use it.