multi_to_single: add expanded argument.
When reading a multi-segment record, the multi_to_single function is
used to stitch the segments together into a virtual single-segment
record.

To enable this to work in expanded (non-smooth-frames) mode, we want
to combine the 'e_p_signal' or 'e_d_signal' arrays from the individual
segments, rather than 'p_signal' or 'd_signal'.