Implement reading signals in format 24.
In signal format 24, each sample is stored in three bytes, as a
little-endian signed integer.  Since most architectures don't natively
support 24-bit integers, neither does numpy, and therefore in order to
read the signal, it must be converted from the on-disk format into
some format that numpy can work with (i.e., 32-bit signed integers.)

Previously, when loading a format-24 record using rdrecord or rdsamp,
the internal function _rd_dat_file made a halfhearted attempt to read
the signals, by extracting the two most significant bytes of each
sample, discarding the least significant byte, and storing the result
as a 16-bit integer.  This meant not only that the precision was
silently reduced, but the resulting physical sample values were
incorrect by a factor of 256.

(Additionally, the implementation of _rd_dat_file was broken if the
signal file contained a prolog, if the signal file could not be
mmapped, or if the entire signal file was not meant to be read at
once.)

Fix this by handling format 24 in the same way as the other unaligned
formats (212, 310, and 311):

 - Read the data initially as an array of unsigned bytes; this means
   the 'dtype' passed to numpy.fromfile is numpy.dtype('<u1') and the
   'count' is the number of bytes rather than the number of samples.

 - Reformat the array of bytes into an array of integers in the
   _blocks_to_samples function.  This means format 24 must be
   considered an "unaligned" rather than an "aligned" format.

In order to avoid making unnecessary copies of the data, rather than
using numpy arithmetic operations, the middle and low bytes are
*copied* from the input array into the corresponding locations in an
8-bit view of the 32-bit output array.  This is dependent on the
system byte order.