Remove obsolete and broken logic for handling 24-bit input.
Previously, the data type string for signal format 24
(DATA_LOAD_TYPES['24']) was set to '<i3', which is numpy shorthand for
"little-endian, signed integers, three bytes per element", although
numpy does not actually support such a format.

This file format was previously accommodated by special cases in
_rd_dat_file (for local files) and in _stream_dat (for remote files).
Both of these implementations were broken, in that they would silently
discard the least significant byte and return a 16-bit integer array.
The implementation for local files was further broken in that it only
worked if the input file was mmappable, if the range to be retrieved
was the entire input file, and if the file consisted of an even number
of samples.

Format 24 is now accommodated by reading the data as unsigned 8-bit
integers (DATA_LOAD_TYPES['24'] = '<u1') and reformatting the array
afterwards (in _blocks_to_samples), as was previously done for formats
212, 310, and 311.  This makes the previous logic in _rd_dat_file and
_stream_dat unnecessary, so remove it.