wfdb.io.download: add _get_url function.
This function provides a simple wrapper for requests.get, handling the
common case of reading the entire contents of a binary file and
raising an exception if the file doesn't exist.  This function is a
temporary shim to assist in porting the wfdb.io.download functions to
using wfdb.io._url.