_rd_dat_signals: optimize construction of expanded arrays.
When reading signal data with smooth_frames = False, we want to return
a one-dimensional array of samples for each signal, which means
extracting a range of samples from each frame and concatenating those
ranges into a single numpy array.

Previously this was done by constructing an array of all the desired
sample indices, and using that array as a subscript into the input
sig_data array.  This is inefficient in both time and memory.

Instead, extract the desired samples as a slice of a two-dimensional
array, and reshape that slice to obtain a 1D array.

(Using reshape(-1) instead of flatten() means that if samps_per_frame
is 1, the array does not need to be copied.  If the array does need to
be copied, the reshape operation is still faster than any Python loop
and doesn't require additional memory.)