rdrecord: correctly infer length for multi-frequency records.
If the header file does not specify the record length, it must be
inferred from the size of the signal file (dividing the data length by
the number of samples per frame.)

Previously, this calculation assumed that tsamps_per_frame == n_sig,
which is not necessarily the case.  Correctly calculate the total spf
for the first signal file and use that as the divisor, rather than the
number of signals.  Rename the "n_sig" parameter of _infer_sig_len to
tsamps_per_frame accordingly.