rdrecord: correctly infer length for signal files with a prolog.
If the header file does not specify the record length, it must be
inferred from the size of the signal file (dividing the data length by
the number of samples per frame.)

Previously, this calculation assumed that the file did not contain a
prolog.  Correctly calculate the data length by subtracting the prolog
length (byte offset) from the size of the file.