RangeTransfer: fix exception handling.
If a normal exception occurs while reading an HTTP response, we want
to read the remaining data so that the connection can be reused for
another transfer.

If the RangeTransfer is deleted without calling close() or __exit__(),
the response object must still be explicitly closed so that it no
longer counts against the connection pool limit.  This doesn't happen
automatically when a Response is garbage-collected, which is probably
a bug in python-requests.