Parse header lines starting at the beginning of the line.
When reading a header file, each line is parsed by matching it against
a regular expression.  Previously, this was done by searching for the
pattern *anywhere* in the input line (using findall()) rather than at
the start of the line (using match()).

In particular, if the first token on the line (e.g., a signal file
name or segment name) contains a disallowed character, this must be
treated as an error, rather than silently ignoring everything up to
and including the disallowed character.

Note, however, that horizontal whitespace is permitted at the start of
the line, so the expressions must begin with "[ \t]*".