rdann: handle multiple consecutive SKIPs.
In WFDB-format annotation files, annotation timestamps are represented
as an offset from the previous annotation.  When this offset is less
than 0 or greater than 1023, a SKIP pseudo-annotation is used; when
the offset is greater than 2**31 - 1 or less than -2**31, multiple
SKIPs must be used.  Thus, proc_core_fields must be able to handle an
arbitrary number of SKIPs in a row, preceding the actual annotation,
and add all of the offsets together to obtain the final timestamp.