Add a test case for all WFDB binary signal formats.
The record "binformats" contains one signal in each of the ten WFDB
binary formats (8, 16, 61, 80, 160, 212, 310, 311, 24, and 32.)  In
this record, sample j of signal i is equal to:

  (i + 16843019 * j) % ((1 << adcres) - 1) + 1 - (1 << (adcres - 1)))

Note that the length of the record is 499 samples, so each of the
bit-packed data files ends with an incomplete data block.

Use this record to test that it is possible to read all of the formats
correctly, including when we skip one or two samples from the start
and/or end of the record.  (Skipping samples is expected to give
incorrect results for format 8, so that signal is not required to
match.)

We do not test writing, since not all formats are currently supported
by wr_dat_file.