convert_dtype: don't change p_signal when smooth_frames is false.
If physical is true, and smooth_frames is false, and e_p_signal is
None, this function should raise an exception.  It should not silently
modify p_signal instead.

The previous code didn't make any sense: p_signal is a numpy array, so
you can't change its data type by assigning to a slice of it, and even
if that conversion were possible, the slicing was being done along the
wrong axis.  This reverts a dubious change in commit
725e7a65f90f5e86620b902f4a5ad0734f1bec46, which seems unrelated to the
rest of that commit.