Record.__eq__: replace exact type comparisons with isinstance.
When comparing the attributes of two records, if a value is an
instance of a class derived from 'np.ndarray' or a class derived from
'list', it should be treated the same way as an 'np.ndarray' or a
'list'.  Note that there currently are no such classes defined or used
by this package.

Note also that in any case, the respective types of the two attributes
must be equal in order for the values to be considered equal.