Add example annotation files for a multi-frequency record.
The annotation files given here may be useful as examples and for
testing.  They were generated by the following commands (WFDB 10.6.2):

    gqrs -r 03700181 -o gqrsl

(an annotation file with no explicit time resolution - meaning the
resolution is assumed to be one tick per WFDB frame, i.e. 1/125 s.)

    gqrs -H -r 03700181 -o gqrsh

(an annotation file with a time resolution of 1/500 s, which, in this
case, matches the sampling interval of the signal in question.)

    sqrs -r 03700181 && mv 03700181.qrs 03700181.sqrs

(an annotation file with a time resolution of 1/250 s, which doesn't
match either the frame interval or the sampling interval.)