plot_items: add sampling_freq and ann_freq parameters.
An annotation file (represented by a wfdb.Annotation object) may have
a "sampling frequency" that differs from the sampling frequency or
frequencies of the signals themselves.  This will be the case, for
example, for annotations generated by programs like sqrs that operate
on an upsampled or downsampled copy of the input signals.

In any event, when plotting annotations, we need to translate the
annotation time into a sample number in order to display it in the
correct location.

Furthermore, in the future, we want to permit plotting multiple
synchronized signals that are sampled at different frequencies.

Therefore, to disambiguate between the many possible "sampling
frequencies" invvolved, add parameters 'sampling_freq' and 'ann_freq'.
Both parameters are optional and default to the value of 'fs'.  Either
may be a list (one entry per channel), allowing the API to accomodate
multi-frequency data.  Currently, if 'sampling_freq' is a list, all of
its elements must be equal.