plot_wfdb: handle differing frame/annotation frequencies.
An annotation file (represented by a wfdb.Annotation object) may have
a "sampling frequency" that differs from the sampling frequency or
frequencies of the signals themselves.  This will be the case, for
example, for annotations generated by programs like sqrs that operate
on an upsampled or downsampled copy of the input signals.

To handle this case, since record.fs does not equal annotation.fs, we
must explicitly specify both sampling_freq and ann_freq when calling
plot_items.