plot_signal: accept non-smooth signal data as input.
When plotting signals, in addition to allowing the signal argument to
be a one-dimensional or two-dimensional array, allow it to be a list
of one-dimensional arrays (so that each channel can have a different
length.)

This means that the array of X values (t) must be calculated
separately for each channel, as both the sampling frequency and the
number of samples may vary.  (Here, we don't require that the lengths
of the signals, in seconds, must all be the same, although they always
will be when plotting a WFDB record.)

The sig_len parameter makes no sense if the channels have different
lengths; this parameter is now unused, and marked as deprecated.