get_plot_dims: accept non-smooth signal data as input.
When plotting signals, in addition to allowing the signal argument to
be a one-dimensional or two-dimensional array, allow it to be a list
of one-dimensional arrays (so that each channel can have a different
length.)

Using _expand_channels here, as in plot_annotation, allows the logic
to be simplified.

The sig_len return value makes no sense if the channels have different
lengths; this return value is now marked as deprecated, and will be
set to None if the channel lengths differ.