plot_wfdb: plot non-smooth records.
If a record is loaded in "non-smooth" ("expanded") mode, it can
contain signals of different lengths sampled at different frequencies.
In such a case, we want to plot each signal at its original
frequency in order to see the effect of the sampling and the temporal
relationships between the signals.

plot_items now allows the signal argument to be either a numpy array
(as in p_signal or d_signal, loaded when 'smooth_frames=True') or a
list of arrays (as in e_p_signal or e_d_signal, loaded when
'smooth_frames=False'); in the latter case, plot_wfdb has to calculate
and provide the correct per-channel sampling frequencies.

(The annotation file, if any, may have its own sampling frequency,
which may differ from the signal sampling frequencies.  The *default*,
if the annotation file doesn't specify a frequency, is always the
record frame frequency (fs), not the sampling frequency of any
particular signal.)