Remove requirements.txt.
Having the file 'requirements.txt' implies that there is a particular
set of external package versions that are recommended, if not
required.

That's not true; all of the packages that wfdb-python requires are
quite stable and any vaguely recent version will do, and you should
generally be using the most up-to-date version that is compatible with
your application requirements.

Moreover, there's no specific set of external package version numbers
that works across all platforms we might want to test.

Therefore, remove the file requirements.txt, and for purposes of
*testing*, use whatever pip selects by default.