pyproject.toml: relax requirement for matplotlib to 3.2.2.
As of 2022-05-26, a certain popular cloud computing service comes with
matplotlib 3.2.2 preinstalled, along with other packages that will
cause a conflict if you try to install the latest version of
matplotlib (3.5.2).

I'm not aware of any incompatibilities with this version, so relax
wfdb's requirement in order to make installation smoother.