BaseRecord: always define the base_datetime attribute.
Previously, if a record contained both a base date and base time, then
rdrecord would set the undocumented 'base_datetime' attribute
accordingly.

This attribute is useful to have in general, so define it for all
BaseRecords, Records, and MultiRecords.  It is defined as a property
that aliases the base_date and base_time attributes.  That is to say,
if the package or application sets record.base_date and
record.base_time, then record.base_datetime is calculated
automatically; if the package or application sets
record.base_datetime, then record.base_date and record.base_time are
calculated automatically.

Internally, we continue to store the time and date attributes
separately, to accommodate records where the base time is defined but
the base date is not.

Note that when storing a value in record.base_datetime, it must be a
naive datetime object.  Records don't have a time zone, and therefore
trying to set the base time to an aware datetime object is incorrect.