_adjust_datetime: simplify calculation of base_datetime.
_adjust_datetime is called by _arrange_fields, which is called by
rdrecord, in order to calculate the base date and time, given that a
subset of the record has been selected (using sampfrom).

Previously, this function would calculate base_datetime by combining
base_date and base_time; this is now unnecessary because base_datetime
is set implicitly when base_date and base_time are set.  Likewise, it
is unnecessary to set base_time and base_date afterwards, because
these are set implicitly when base_datetime is set.