_rd_compressed_file: read signals in 1048576-sample chunks.
A bug in libsndfile [1] means that if we try to read a chunk of more
than 16777216 (2**24) total samples (across all channels) from a FLAC
file, and the number of channels is not a power of two, then
sf_read_short will return zero.

Work around this bug by allocating an array ourselves and reading a
small block of samples at a time (a FLAC file has at most 8 channels,
at most 7 if not a power of two, and 7 * 1048576 < 16777216).

[1] https://github.com/libsndfile/libsndfile/issues/431