plot_wfdb: add sharex argument.
This argument allows the caller to specify whether the horizontal axes
should be synchronized across all channels (subplots) in a plot.

In general, this behavior is desirable as it lets the viewer see the
temporal relationship between channels.  However, in the case where
the record is multi-frequency *and* we are displaying sample numbers
on the horizontal axis, matplotlib unfortunately does not appear to
have a way to synchronize the axes automatically.

Therefore, in the default "auto" case, enable X sharing as long as the
time units for all channels are the same.