Record.adc: round samples rather than truncating.
When converting physical to digital values, the input values are
floating-point and therefore may have rounding errors.  We want to
round each value to the nearest integer before calling istype, in
order to avoid adding a bias towards zero.

This applies to adc in all four modes: expanded and not expanded,
in-place and not-in-place.

("In-place" here means both that we can overwrite the original
floating-point arrays, and that the record attributes will be updated
afterwards.  In "not-in-place" mode, we are operating on copies of the
original arrays.)