TestSignal: Test physical-to-digital conversion.
When converting physical to digital values, the input values are
floating-point and therefore may have rounding errors.  We want to
ensure that the package rounds each value to the nearest integer, in
order to avoid adding a bias towards zero.

This test case tests all four code paths in Record.adc, and also
checks that the high-level wfdb.wrsamp function works correctly for a
round trip from physical to digital to physical.  Note that wrsamp
currently doesn't support expanded/multi-frequency data, so that is
not tested.