wr_dat_file: do not modify contents of the d_signal argument.
When writing to an output file, it may be necessary to transform the
data in various ways, which is often most efficient to do by modifying
the array in-place.

However, for API cleanliness, the wr_dat_file function shouldn't
modify its argument.  Instead, the function itself should make a copy
of the array if necessary.  (In the future, we may ultimately be able
to avoid a copy if the array is already in the correct format.)