wr_dat_file: support FLAC output formats.
Writing FLAC formats, like reading FLAC formats, is done using the
soundfile package, which is currently treated as an optional
dependency (since on some systems it requires an external copy of
libsndfile.)

This code performs, in effect, the inverse of _rd_compressed_file.  As
with _rd_compressed_file, this has room to be optimized somewhat.