Record.wrsamp: ignore samps_per_frame when expanded is false.
When the expanded argument to wrsamp is false, we are writing the
d_signal data, which contains one sample per frame by definition.
Therefore, the header file must not claim that the signal files will
contain multiple samples per frame.

The samps_per_frame attribute of the Record object may or may not be
set appropriately for the actual data contained in the Record object;
for example, rdrecord(r, smooth_frames=True) will set samps_per_frame
according to the number of samples per frame in the original record.