run-tests.yml: check that format is acceptable to black.
If all code is supposed to be formatted according to black, we want to
check this and fail if any code is not formatted correctly.

Simply running "black ." here doesn't help, as that simply rewrites
any incorrectly-formatted files and exits with status 0.  Instead, run
with --check (fail if files are wrongly formatted) and --diff (print
differences to standard output).