Fix type inference for ALLOWED_TYPES.
When using mypy to check the package, it will attempt to infer types
that are not specified.

Currently, mypy is able to understand that constructing a dict from a
Sequence[Tuple[X, Y]] yields a Dict[X, Y], but if we use lists instead
of tuples, mypy doesn't understand and seems to think the result is a
Dict[X, X].

This is probably a bug in mypy, but in any case, using tuples here
rather than lists is more idiomatic and doesn't affect the behavior of
the code.