MultiHeaderMixin: declare types of object attributes.
mypy (and likely other static analysis tools) will complain if an
object attribute is used without being defined either in the class or
the class's __init__ method.

The MultiHeaderMixin class is not public and not meant to be
instantiated at all, but refers to attributes that are defined by the
MultiRecord class which inherits from it.  Declare the types of these
attributes here so that mypy knows they exist.