_parse_record_line: annotate type of record_fields variable.
This variable is used in a strange way; initially it contains string
values extracted from the record line, then later these string values
are replaced with appropriately-typed values for each field.  However,
there's no real way to make this type-safe without changing the API of
the function.

Annotate the variable's type so that mypy will not treat it as a
Dict[str, Optional[str]].