adc (inplace=True, expanded=False): correctly handle NaNs.
When converting physical to digital sample arrays, we must replace NaN
values (which represent a missing sample) with the appropriate
invalid-sample sentinel value.

This is done correctly for normal uses of the package, but if the
application directly invoked adc(inplace=True), NaNs would not have
been handled (and were instead set to an implementation-defined
value.)

(Note that we don't use inplace=True internally because this
overwrites the original floating-point array.  Applications may want
to use inplace=True to save memory, but this requires knowing that the
original array is no longer needed.)