README: remove confusing comments about poetry.
poetry is used as a build tool, but people who want to install and use
the package have no need to interact directly with poetry.  Moreover,
the "poetry install" command doesn't work.  Remove comments that
suggest otherwise.

(If people are already using poetry for reasons of their own, there's
nothing stopping them from adding wfdb as a dependency to their own
project.  In that case, they should know to consult the documentation
of the tools they're using.)

Also clarify that installing "dev" dependencies is an optional thing
for developers, and not a required installation step.